Add docs publishing CI job and integrate tutorials and use new theme (#57)
* Add docs publishing CI job and integrate tutorials and use new theme

This commit adds a new CI job to build and publish the documentation to
https://qiskit.org/documentation/optimization at release time. As part
of this change it is necessary to integrate the tuorials in the
documentation buid. This commit does this so a docs build always
includes the tutorials. This also switches the docs theme to be the new
theme to be consistent with the rest of qiskit's documentation.

* fix actions env

* fix copyright

* add fetch depth to tutorials job

* fix tutorials path

* fix notebook image

Co-authored-by: Manoel Marques <manoel.marques@ibm.com>