Bump reno requirement to latest version and set default_branch (#220)
When we created a stable branch for qiskit-optimization we encountered
a bug in reno that required a branch named master to exist for traversing
the git history. To workaround that issue we temporarily added back a
branch named master while we waited for the fix [1] to be included in a
release. A new reno release has been published [2] and we can rely on
that version and delete the idle master branch now. This commit makes
this change and switches to using reno>=3.4.0 and uses the new
default_branch configuration option added in 3.4.0 to fix the issue.

[1] https://opendev.org/openstack/reno/commit/ed6bbae82e01edf781534d9e9cce21e0c55a49a6
[2] https://pypi.org/project/reno/3.4.0/

(cherry picked from commit 4953e56e3c062df42e7e51fa6b89172e1f69e3c1)