Duck-type VQE interim results
Add a check for the type of data expected in the callback wraper in VQEProgram.

This will allow us to close Qiskit-Partners/qiskit-runtime#38

Anticipating changes in the Qiskit Runtime API which will start emitting final results in the WebSocket stream. When that happens, if we don't duck-type in the VQE callback wrapper receiving the interim results, it will get a ValueError exception.

Already done in Qiskit/qiskit-nature#350