Create bin_packing.py (#258)
* Create bin_packing.py

Adding the bin packing problem to the applications on qiskit_optimization

* Improving Bin Packing problem and adding test

I followed the directions from Takashi to make the code neat, adding a max_number_bins and a test for the bin packing problem.

* Solving some mistakes with the bin packing test

* following some suggestions of woodsp-ibm

* Interpret reformatting

Following the suggestions from woodsp-ibm and a-matsuo I changed the interpret format to give the results as a list of lists where each sublist is the items of each bin

* Testing max_number_of_bins

Improving the test of max_number_of_bins to test its quadratic program version and removing the max_weight test

* Update qiskit_optimization/applications/bin_packing.py

Co-authored-by: Steve Wood <40241007+woodsp-ibm@users.noreply.github.com>

* Following suggestions from Takashi

I followed some suggestions from @t-imamichi about following the same notation from wikipedia.

* Update qiskit_optimization/applications/bin_packing.py

Co-authored-by: Takashi Imamichi <31178928+t-imamichi@users.noreply.github.com>

* Update qiskit_optimization/applications/bin_packing.py

Co-authored-by: Takashi Imamichi <31178928+t-imamichi@users.noreply.github.com>

* Update __init__.py

Adding bin_packing to __init__.py

* Adding a reno note for the BinPacking problem

Adding a reno note for the bin packing problem

* adding a release note

Adding a release note and solve a minor mistake.

* Adding a plot feature

Adding a plot for the visualization of the solution.

* Update qiskit_optimization/applications/__init__.py

Co-authored-by: Takashi Imamichi <31178928+t-imamichi@users.noreply.github.com>

* Update draw function

Adding suggestions from @t-imamichi about the visualization of the results.  Testing if the library matplotlib exists. I used the same format as graph_optimization_application

* Update bin_packing.py

* Following suggestions about Figure

Following the suggestions from @woodsp-ibm about an empty class if matplotlib is not installed and the name of the function.

* Update bin_packing.py

* Update qiskit_optimization/applications/bin_packing.py

Co-authored-by: Steve Wood <40241007+woodsp-ibm@users.noreply.github.com>

* Adding a test for get_figure

I add a test for the visualization of the solution of the bin packing problem.

* Update test_bin_packing.py

Adding suggestions from @t-imamichi  about adding the test when there is no matplotlib.

* Skip unit test if matplotlib not installed

Co-authored-by: Takashi Imamichi <31178928+t-imamichi@users.noreply.github.com>
Co-authored-by: Manoel Marques <Manoel.Marques@ibm.com>
Co-authored-by: Steve Wood <40241007+woodsp-ibm@users.noreply.github.com>