Fix code example in ``RecursiveMinimumEigenOptimizer`` (#357)
* Fix code example in ``RecursiveMinimumEigenOptimizer``

The code example was passing a ``QAOA`` as ``optimizer``
but this needs to be an ``OptimizationAlgorithm``.

Also, the formatting for the reference was adjusted. Using a colon
starts a new subsection in Sphinx, which we don't want here.

* copyright

(cherry picked from commit 1f74e426a532ae4ab09a80c213b97a8e3d38f3d4)